server/requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
django-cors-headers
django-debug-toolbar==3.1.1
django-filter==2.4.0
django-versatileimagefield==2.0
djangorestframework-simplejwt==4.4.0
drf-flex-fields==0.8.6
importlib-metadata==2.0.0
Markdown==3.3
Pillow==8.0.0
PyJWT==1.7.1
python-magic==0.4.15
djongo==1.3.6
pymongo==3.12.0
pytz==2021.1
six==1.15.0
sqlparse==0.2.4
typing==3.7.4.3
zipp==3.3.0
django-cors-headers